kotlin/journal3: Move classes to their appropriate package
Move filesystem-related test classes into their appropriate package.
